lab1/pyproject.toml
[tool.poetry]
name = "assignment-visual"
version = "0.1.0"
description = "Visual Learning Lab 1"
authors = ["Owen <oyeonk@andrew.cmu.edu>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
absl-py = "==1.4.0"
aiohttp = "*"
aiosignal = "*"
appdirs = "==1.4.4"
asynctest = "==0.13.0"
async-timeout = "*"
attrs = "*"
backcall = "*"
blinker = "*"
brotlipy = "==0.7.0"
cachetools = "==5.3.0"
certifi = "==2022.12.7"
cffi = "*"
charset-normalizer = "==3.0.1"
click = "*"
colorama = "*"
cryptography = "*"
cycler = "*"
debugpy = "*"
decorator = "*"
docker-pycreds = "*"
entrypoints = "*"
flit_core = "*"
frozenlist = "*"
gitdb = "*"
GitPython = "*"
grpcio = "==1.51.1"
idna = "==3.4"
imageio = "==2.25.0"
importlib-metadata = "==6.0.0"
ipykernel = "*"
ipython = "*"
jedi = "*"
joblib = "==1.2.0"
jupyter_client = "*"
jupyter_core = "*"
kiwisolver = "*"
Markdown = "==3.4.1"
MarkupSafe = "==2.1.2"
matplotlib = "^3.9.2"
matplotlib-inline = "*"
mkl-fft = "^1.3.1"
mkl-random = "*"
mkl-service = "==2.4.0"
multidict = "*"
nest-asyncio = "*"
numpy = "^1.24.0"
oauthlib = "*"
olefile = "*"
packaging = "*"
parso = "*"
pathtools = "*"
pexpect = "*"
pickleshare = "*"
Pillow = "==9.4.0"
prompt-toolkit = "*"
protobuf = "==3.20.3"
psutil = "*"
ptyprocess = "*"
pyasn1 = "==0.4.8"
pyasn1-modules = "==0.2.8"
pycparser = "*"
Pygments = "*"
PyJWT = "*"
pyOpenSSL = "*"
pyparsing = "*"
PySocks = "*"
python-dateutil = "*"
pyu2f = "*"
PyYAML = "*"
pyzmq = "*"
requests = "==2.28.2"
requests-oauthlib = "==1.3.1"
rsa = "*"
scikit-learn = "*"
scipy = "*"
setproctitle = "*"
six = "==1.16.0"
smmap = "*"
tensorboard-plugin-wit = "==1.8.1"
threadpoolctl = "==3.1.0"
torch = { url = "https://download.pytorch.org/whl/cu113/torch-1.12.1%2Bcu113-cp310-cp310-linux_x86_64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/cu113/torchaudio-0.12.1%2Bcu113-cp310-cp310-linux_x86_64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu113/torchvision-0.13.1%2Bcu113-cp310-cp310-linux_x86_64.whl" }
tornado = "*"
tqdm = "*"
traitlets = "*"
typing_extensions = "==4.4.0"
urllib3 = "==1.26.14"
wandb = "*"
wcwidth = "*"
Werkzeug = "==2.2.2"
wget = "==3.2"
yarl = "*"
zipp = "==3.12.1"
opencv-python = "^4.10.0.84"  # Added opencv-python
tk = "^0.1.0"
torchmetrics = "^1.4.2"
pycocotools = "^2.0.8"
tensorboard = "^2.17.0"
tensorflow = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"